+++ achives/deploy/AvaxUsdtStrategy.ts
@@ -43,12 +43,8 @@ const deployFunction: DeployFunction = async function (hre: HardhatRuntimeEnviro
 
   if (network.name !== "hardhat") {
     const AVAXUSDTStrategy = await ethers.getContract<LPStrategy>("AVAXUSDTStrategy");
-    if ((await AVAXUSDTStrategy.feeCollector()) !== xMerlin) {
-      await AVAXUSDTStrategy.setFeeCollector(xMerlin); // Temporary until changed to correct feeCollector
-    }
-    if ((await AVAXUSDTStrategy.owner()) !== xMerlin) {
-      await AVAXUSDTStrategy.transferOwnership(xMerlin);
-    }
+    await AVAXUSDTStrategy.transferOwnership(xMerlin);
+    await AVAXUSDTStrategy.setFeeCollector(xMerlin); // Temporary until changed to correct feeCollector
   }
 };
